+++ Tests/TuistGeneratorTests/Generator/ProjectFileElementsTests.swift
@@ -792,7 +792,7 @@ final class ProjectFileElementsTests: TuistUnitTestCase {
         )
 
         // Then
-        XCTAssertEqual(groups.compiled.flattenedChildren, [
+        XCTAssertEqual(groups.frameworks.flattenedChildren, [
             "ARKit.framework",
         ])
 
@@ -840,7 +840,7 @@ final class ProjectFileElementsTests: TuistUnitTestCase {
         )
 
         // Then
-        XCTAssertEqual(groups.compiled.flattenedChildren, [
+        XCTAssertEqual(groups.frameworks.flattenedChildren, [
             "Test.framework",
         ])